refactor: YAML handler for configs and file structure
