+++ display/app/api/config/route.ts
@@ -1,0 +1,6 @@
+import { getRawConfig } from '@/lib/server/configLoader';
+import { NextResponse } from 'next/server';
+
+export function GET() {
+  return NextResponse.json(getRawConfig());
+}
