+++ display/app/api/overlay/subscribe/route.ts
@@ -1,7 +1,7 @@
 export const dynamic = "force-dynamic"; 
 export const fetchCache = "force-no-store";
 
-import { getOverlayData } from '@/lib/overlay/overlayInfo';
+import { getOverlayData } from '@/lib/server/overlayHandler';
 import { newStream } from '@/lib/transmitter/helper';
 
 export async function GET() {
