+++ display/lib/helper/config.js
@@ -1,0 +1,8 @@
+
+const URL = 'http://localhost:3000'
+
+// Configurations
+export async function getConfig() {
+    const res = await fetch(`${URL}/api/config`);
+    return res.json();
+}
