+++ display/lib/server/overlayHandler.js
